refactor(cli): split tsserver definition handlers
